requirements.txt
opencv-python
numpy
pytesseract
pillow
google-generativeai
